refactor: share RoPE tables and rotation via a single rope.py
apply_rope and the cos/sin table construction were duplicated with
identical math in layers.py (RotaryAttention) and plan_a_model.py
(_rope_tables) — two copies to keep in sync. Both now import from a
tiny config-free rope.py, so plan_a_model stays fully parametrized.
The attention/Block classes themselves stay separate on purpose: they
differ for real reasons (GQA + cross-attention vs simple causal MHA,
different hidden-dim rounding) and merging them would change param
trees and break checkpoint compatibility.

Numerics are unchanged: golden-run and reference-numerics tests pass.
