+++ plan_a_model.py
@@ -6,7 +6,7 @@ only to positions <= t on every iteration, so depth refines the prediction with 
 future-token leak (contrast docs/findings/2026-06-13-cross-window-hunch-inert.md,
 where the loop could only reach the next window and the gradient killed it).
 
-Self-contained and fully parametrized (dim, vocab, heads, depth, seq) so the exact
+Config-free and fully parametrized (dim, vocab, heads, depth, seq) so the exact
 same architecture runs at tiny toy-task scale and at real scale — the ablation
 harness tests the thing we'd ship, not a stand-in. See docs/design/plan-a.md.
 """
@@ -15,17 +15,7 @@ import jax
 import jax.numpy as jnp
 from flax import nnx
 
-
-def _rope_tables(max_pos, head_dim):
-    inv_freq = 1.0 / (10000 ** (jnp.arange(0, head_dim, 2) / head_dim))
-    freqs = jnp.outer(jnp.arange(max_pos), inv_freq)
-    return jnp.cos(freqs), jnp.sin(freqs)
-
-
-def apply_rope(x, cos, sin):
-    x1, x2 = jnp.split(x, 2, axis=-1)
-    rotated = jax.lax.complex(x1, x2) * jax.lax.complex(cos, sin)
-    return jnp.concatenate([rotated.real, rotated.imag], axis=-1).astype(x.dtype)
+from rope import rope_tables, apply_rope
 
 
 class CausalAttention(nnx.Module):
@@ -42,7 +32,7 @@ class CausalAttention(nnx.Module):
         self.o = nnx.Linear(dim, dim, rngs=rngs, dtype=dtype)
         self.q_norm = nnx.RMSNorm(self.head_dim, epsilon=1e-6, rngs=rngs, dtype=jnp.float32)
         self.k_norm = nnx.RMSNorm(self.head_dim, epsilon=1e-6, rngs=rngs, dtype=jnp.float32)
-        cos, sin = _rope_tables(max_pos, self.head_dim)
+        cos, sin = rope_tables(max_pos, self.head_dim)
         self.cos, self.sin = cos, sin
 
     def __call__(self, x, pad_bias=None):
